index: deserialize into structure
Until now, the index was only encoded on checkout. This adds the
converse operation to decode the index into a struct, as well as a
testcase to validate the round-trip behavior.

In order to support more complete serialization, there is the bare
minimum support for parsing git extensions. Namely, the contents
remain unused. The only thing we validate for extensions are the
signature and expected length.